+++ packages/uni-app-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @js-css/uni-app-react
 
+## 0.0.12
+
+### Patch Changes
+
+- fix: app document missing unmount & fix after&before css side effect
+  feat: app event support stop event
+  todo: app event side effect origin vue event system
+
 ## 0.0.11
 
 ### Patch Changes